Extract a CertificateAuthority class
To encapsulate generating new signed certificates
